Condition now features a requires_stop() method (cf. issue #39).
SoftwareCondition now longer checks whether or not the callback can be evaluated in the build_sequence_* methods. Its requires_stop returns an according value. Handling this is responsibility of Sequencer.
SoftwareCondition now also passes the current number of completed iterations to the callback for use in loops. If the Condition is used by a branch, this value will always be zero.