Safer dealing with duplicate ids when storing pulses.
JSONEncoder previously did not check whether subpulses are identical to already known pulses with the same id but just replaced them with a reference to the already known subpulse in the serialization. This could lead to unexpected behavior. JSONEncoder now raises a RuntimeError instead in this case.
PulseStorage adopts a transaction model and takes care that in case of such an exception, a pulse is not partially serialized, i.e., if a pulse serialization aborts due to a duplicated subpulse identifier, no changes to the storage is made (that is, new subpulses that were already succesfully serialized will be discarded if an error is encountered for another subpulse).