Extended tests for _internal_create_program to check that the result is the same as doing old Sequencing and constructing MultiChannelProgram from resulting InstructionBlock.
Fixed behavior of AtomicPulseTemplate (and DummyPulseTemplate). Added measurements to compare_key of Loop.
Differences in behavior as detected by those tests are left as comments and will be reported to corresponding issue #329 . Fails are currently written not to fail in case of deviation (until a final desired behavior has been determined).