Monkey patching for Transformation.from_pandas().
Moved implementation of the method into a try-except-block outside of the
class and patching it in via monkey patching.
Reason: sphinx documentation building crashed due to unknown name "pandas"
in the type hint if pandas is not imported. Further, this method only
makes sense if the pandas package is available. However, pandas is not a
necessary requirement for anything else in the module. Having a separate
block that only declares the function when the requirement (pandas) is
present seems to be a viable solution.