Attempt to enable namespace dot-notation (e.g. "foo.bar") in sympy expressions.
Created customized version of the auto_symbol token transformation used by sympy.parse_expr and mocked that into sympy when calling sympify and co.
Works well for sympify but not at all for lambdified and compiled expressions.
An issue for lamdified expression is that the namespaced parameters will become kwargs and python cannot deal with the dot in there.
anything to be done? WIP? probably just scrap this..